+++ dev-packages/application-package/package.json
@@ -34,10 +34,10 @@
     "@types/write-json-file": "^2.2.1",
     "changes-stream": "^2.2.0",
     "fs-extra": "^4.0.2",
-    "is-electron": "^2.1.0",
-    "request": "^2.82.0",
+    "is-electron": "^2.2.0",
+    "request": "^2.88.0",
     "semver": "^5.4.1",
-    "write-json-file": "^2.2.0"
+    "write-json-file": "^4.2.1"
   },
   "devDependencies": {
     "@theia/ext-scripts": "^0.14.0"
