+++ dev-packages/cli/package.json
@@ -35,16 +35,14 @@
     "@types/chai": "^4.2.7",
     "@types/mkdirp": "^0.5.2",
     "@types/mocha": "^5.2.7",
-    "@types/puppeteer": "^2.0.0",
-    "@types/requestretry": "^1.12.3",
+    "@types/requestretry": "^1.12.5",
     "@types/tar": "^4.0.3",
     "chai": "^4.2.0",
-    "mkdirp": "^0.5.0",
+    "mkdirp": "^0.5.1",
     "mocha": "^7.0.0",
-    "puppeteer": "^2.0.0",
-    "puppeteer-to-istanbul": "^1.2.2",
-    "requestretry": "^3.1.0",
-    "tar": "^4.0.0",
+    "playwright": "^0.9.21",
+    "requestretry": "^4.1.0",
+    "tar": "^5.0.5",
     "unzip-stream": "^0.3.0",
     "yargs": "^11.1.0"
   }
