+++ dev-packages/cli/src/run-test.js
@@ -16,31 +16,14 @@
 
 // tslint:disable:no-any
 
-import * as net from 'net';
-import * as puppeteer from 'puppeteer';
-const collectFiles: (options: TestFileOptions) => string[] = require('mocha/lib/cli/collect-files');
+import * as playwright from 'playwright';
+const collectFiles = require('mocha/lib/cli/collect-files');
 
-export interface TestFileOptions {
-    ignore: string[]
-    extension: string[]
-    file: string[]
-    recursive: boolean
-    sort: boolean
-    spec: string[]
-}
-
-export interface TestOptions {
-    start: () => Promise<net.AddressInfo>
-    launch?: puppeteer.LaunchOptions
-    files?: Partial<TestFileOptions>
-    coverage?: boolean
-}
-
-export default async function runTest(options: TestOptions): Promise<void> {
+export default async function runTest(options) {
     const { start, launch } = options;
     const exit = !(launch && launch.devtools);
 
-    const fileOptions: TestFileOptions = {
+    const fileOptions = {
         ignore: options.files && options.files.ignore || [],
         extension: options.files && options.files.extension || [],
         file: options.files && options.files.file || [],
@@ -53,7 +36,7 @@ export default async function runTest(options: TestOptions): Promise<void> {
     collectFiles(fileOptions);
 
     const server = await start();
-    const browser = await puppeteer.launch(launch);
+    const browser = await playwright.launch(launch);
 
     const page = await browser.newPage();
     page.on('dialog', dialog => dialog.dismiss());
@@ -65,11 +48,11 @@ export default async function runTest(options: TestOptions): Promise<void> {
         if (theiaLoaded) {
             return;
         }
-        console.log('loading chai...');
+        console.log('Loading chai...');
         theiaLoaded = true;
         page.addScriptTag({ path: require.resolve('chai/chai.js') });
         page.evaluate(() =>
-            window.addEventListener('beforeunload', () => (window as any)['fireDidUnloadTheia']())
+            window.addEventListener('beforeunload', () => (window)['fireDidUnloadTheia']())
         );
     };
     page.on('frameattached', preLoad);
@@ -78,19 +61,19 @@ export default async function runTest(options: TestOptions): Promise<void> {
     page.on('load', async () => {
         console.log('loading mocha...');
         // replace console.log by theia logger for mocha
-        await page.waitForFunction(() => !!(window as any)['theia']['@theia/core/lib/common/logger'].logger, {
+        await page.waitForFunction(() => !!(window)['theia']['@theia/core/lib/common/logger'].logger, {
             timeout: 30 * 1000
         });
         await page.addScriptTag({ path: require.resolve('mocha/mocha.js') });
-        await page.waitForFunction(() => !!(window as any)['chai'] && !!(window as any)['mocha'] && !!(window as any)['theia'].container, { timeout: 30 * 1000 });
+        await page.waitForFunction(() => !!(window)['chai'] && !!(window)['mocha'] && !!(window)['theia'].container, { timeout: 30 * 1000 });
 
         console.log('loading Theia...');
         await page.evaluate(() => {
-            const { FrontendApplicationStateService } = (window as any)['theia']['@theia/core/lib/browser/frontend-application-state'];
-            const { PreferenceService } = (window as any)['theia']['@theia/core/lib/browser/preferences/preference-service'];
-            const { WorkspaceService } = (window as any)['theia']['@theia/workspace/lib/browser/workspace-service'];
+            const { FrontendApplicationStateService } = (window)['theia']['@theia/core/lib/browser/frontend-application-state'];
+            const { PreferenceService } = (window)['theia']['@theia/core/lib/browser/preferences/preference-service'];
+            const { WorkspaceService } = (window)['theia']['@theia/workspace/lib/browser/workspace-service'];
 
-            const container = (window as any)['theia'].container;
+            const container = (window)['theia'].container;
             const frontendApplicationState = container.get(FrontendApplicationStateService);
             const preferenceService = container.get(PreferenceService);
             const workspaceService = container.get(WorkspaceService);
@@ -105,7 +88,7 @@ export default async function runTest(options: TestOptions): Promise<void> {
         console.log('loading test files...');
         await page.evaluate(() => {
             // replace require to load modules from theia namespace
-            (window as any)['require'] = (moduleName: string) => (window as any)['theia'][moduleName];
+            (window)['require'] = (moduleName) => (window)['theia'][moduleName];
             mocha.setup({
                 reporter: 'spec',
                 ui: 'bdd',
@@ -119,7 +102,7 @@ export default async function runTest(options: TestOptions): Promise<void> {
 
         console.log('running test files...');
         const failures = await page.evaluate(() =>
-            new Promise<number>(resolve => mocha.run(resolve))
+            new Promise(resolve => mocha.run(resolve))
         );
         if (options.coverage) {
             console.log('collecting test coverage...');
@@ -127,7 +110,7 @@ export default async function runTest(options: TestOptions): Promise<void> {
                 page.coverage.stopJSCoverage(),
                 page.coverage.stopCSSCoverage(),
             ]);
-            require('puppeteer-to-istanbul').write([...jsCoverage, ...cssCoverage]);
+            require('playwright-to-istanbul').write([...jsCoverage, ...cssCoverage]);
         }
         if (exit) {
             await page.close();
